lambda/inference/requirements.txt
boto3
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic
